feat: Move tests to their own files and add conftest.py to store test_matrix
